Create tests_utils mongo fn
