+++ lib/src/lib.rs
@@ -7,6 +7,7 @@ pub mod serialization;
 pub mod utils;
 pub mod validate_certificate;
 pub mod wallets;
+pub mod tests_utils;
 
 pub mod initialisation {
     use crate::blockchain_client::BlockchainClientConnection;
