+++ lib/src/merkle_trees/append_only.rs
@@ -186,14 +186,8 @@ where
 mod test {
     use super::*;
     use ark_bn254::Fr as Fr254;
+    use crate::tests_utils::*;
     use ark_std::{rand::Rng, UniformRand};
-    use testcontainers::{
-        core::{IntoContainerPort, WaitFor}, runners::AsyncRunner, ContainerAsync, GenericImage, ImageExt
-    };
-    use tokio::io::AsyncReadExt;
-    use std::time::Duration;
-    use url::Host;
-
     /// makes a vector of n leaves with random values.
     fn make_rnd_leaves<N: UniformRand>(n: usize, mut rng: impl Rng) -> Vec<N> {
         let mut leaves = vec![];
@@ -203,48 +197,47 @@ mod test {
         leaves
     }
 
-    fn get_db_connection_uri(host: Host, port: u16) -> String {
-        format!("mongodb://{}:{}", host, port)
-    }
+    // fn get_db_connection_uri(host: Host, port: u16) -> String {
+    //     format!("mongodb://{}:{}", host, port)
+    // }
 
-    /// This function creates a mongo container and returns it
-    async fn get_mongo() -> ContainerAsync<GenericImage> {
-        GenericImage::new("mongo", "4.4.1-bionic")
+    // /// This function creates a mongo container and returns it
+    // async fn get_mongo() -> ContainerAsync<GenericImage> {
+    //     GenericImage::new("mongo", "4.4.1-bionic")
+    //         .with_exposed_port(27017.tcp())
+    //         .with_wait_for(WaitFor::message_on_stdout("Waiting for connections"))
+    //         .with_startup_timeout(Duration::from_secs(120))
+    //         .start()
+    //         .await
+    //         .unwrap()
+    // }
 
-            .with_exposed_port(27017.tcp())
-            .with_wait_for(WaitFor::message_on_stdout("Waiting for connections"))
-            .with_startup_timeout(Duration::from_secs(120))
-            .start()
-            .await
-            .unwrap()
-    }
+    // /// This function is used to provide a database connection to the tests
+    // async fn get_db_connection(container: &ContainerAsync<GenericImage>) -> mongodb::Client {
+    //     let host = container.get_host().await.unwrap();
+    //     let port = container.get_host_port_ipv4(27017).await.unwrap();
+    //     mongodb::Client::with_uri_str(&get_db_connection_uri(host, port))
+    //         .await
+    //         .expect("Could not create database connection")
+    // }
 
-    /// This function is used to provide a database connection to the tests
-    async fn get_db_connection(container: &ContainerAsync<GenericImage>) -> mongodb::Client {
-        let host = container.get_host().await.unwrap();
-        let port = container.get_host_port_ipv4(27017).await.unwrap();
-        mongodb::Client::with_uri_str(&get_db_connection_uri(host, port))
-            .await
-            .expect("Could not create database connection")
-    }
+    // #[allow(dead_code)]
+    // /// This function is used to print the stdout of a container for test debugging
+    // async fn print_stdout(container: &ContainerAsync<GenericImage>) {
+    //     let mut reader_stdout = container.stdout(false);
+    //     let mut dst = String::new();
+    //     let _ = reader_stdout.read_to_string(&mut dst).await;
+    //     println!("{}", dst);
+    // }
 
-    #[allow(dead_code)]
-    /// This function is used to print the stdout of a container for test debugging
-    async fn print_stdout(container: &ContainerAsync<GenericImage>) {
-        let mut reader_stdout = container.stdout(false);
-        let mut dst = String::new();
-        let _ = reader_stdout.read_to_string(&mut dst).await;
-        println!("{}", dst);
-    }
-
-    #[allow(dead_code)]
-    /// This function is used to print the stderr of a container for test debugging
-    async fn print_stderr(container: &ContainerAsync<GenericImage>) {
-        let mut reader_stderr = container.stderr(false);
-        let mut dst = String::new();
-        let _ = reader_stderr.read_to_string(&mut dst).await;
-        println!("{}", dst);
-    }
+    // #[allow(dead_code)]
+    // /// This function is used to print the stderr of a container for test debugging
+    // async fn print_stderr(container: &ContainerAsync<GenericImage>) {
+    //     let mut reader_stderr = container.stderr(false);
+    //     let mut dst = String::new();
+    //     let _ = reader_stderr.read_to_string(&mut dst).await;
+    //     println!("{}", dst);
+    // }
 
     #[tokio::test]
     async fn test_append_only_tree() {
